requirements.txt
frontmatter
jinja2
python-dotenv 